ingestion-service/requirements.txt
# Core async messaging & logging
aio-pika>=9.0.0
pydantic>=2.0.0
loguru>=0.7.0

# huggingface
sentence-transformers

# LangChain core and integrations
langchain>=0.2.0
langchain-huggingface>=0.0.5  # Or latest available version

# Weaviate client for vector DB
weaviate-client==3.26.7
